Add initial adjusting of rear view for presentation mode
Mimicking Facebook and other apps that use the reveal/slide-out
pattern, this commit doesn't make the rear view take up the full
application frame. Instead, this commit bounds the rear view to the
total visible width, both in partial-reveal and full-presentation mode.

This means none of the rear view's UI will be clipped. Before this
change, if the rear view controller's view had a UISearchBar as a
subview, it would appear cut-off instead of centered within the
revealed view portal.

Here are a couple of demonstrative screenshots:

    http://cl.ly/250o351r3T1q2j0s0S2O

There are some known issues that need to be addressed:

1. The current demo application's rear view controller uses a
   UINavigationBar. Right now, the UINavigationBar's title does not
   animate smoothly, but jumps between its extreme positions when
   entering/exiting presentation mode. I'm not sure what is happening,
   so any feedback would be appreciated.
2. The rear view is not properly resizing itself on orientation
   changes from landscape to portrait.. I'm also not sure what is
   happening here, so any feedback would also be appreciated.