Fixed bug for invalid bounds for presentation mode
Setting a new front view controller while the the PKRevealController is
in presentation mode (either left or right) gives CGRectNull as a new
bound for front container view. Because of this invalid bounds of the
front container view, the animation of leaving the presentation mode
gives black view instead of showing the actual front controller view.